requirements.txt
APScheduler==3.10.4
ccxt==4.4.18
numpy==2.1.3
pandas==2.2.3
python-dotenv==1.0.1
PyYAML==6.0.2
Requests==2.32.3
